requirements.txt
jsonrpclib-pelix==0.3.1
pymongo
pika
requests
redis
lxml
newspaper3k
sklearn
scipy
python-dateutil
pandas
tensorflow==1.15.2
watchdog
